move avro package - rename package in core
