+++ connector/avro/src/test/scala/org/apache/spark/sql/avro/AvroSerdeSuite.scala
@@ -22,6 +22,7 @@ import org.apache.avro.generic.GenericRecordBuilder
 import org.apache.spark.SparkFunSuite
 import org.apache.spark.sql.catalyst.NoopFilters
 import org.apache.spark.sql.catalyst.util.RebaseDateTime.RebaseSpec
+import org.apache.spark.sql.core.avro.{AvroDeserializer, AvroSerializer, IncompatibleSchemaException}
 import org.apache.spark.sql.internal.LegacyBehaviorPolicy.CORRECTED
 import org.apache.spark.sql.types.{IntegerType, StructType}
