+++ sql/core/src/main/java/org/apache/spark/sql/core/avro/AvroCompressionCodec.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.spark.sql.avro;
+package org.apache.spark.sql.core.avro;
 
 import java.util.Arrays;
 import java.util.Locale;
