+++ sql/core/src/main/scala/org/apache/spark/sql/core/avro/AvroOutputWriter.scala
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.spark.sql.avro
+package org.apache.spark.sql.core.avro
 
 import java.io.{IOException, OutputStream}
