+++ sql/core/src/main/scala/org/apache/spark/sql/core/avro/AvroOutputWriterFactory.scala
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.spark.sql.avro
+package org.apache.spark.sql.core.avro
 
 import org.apache.avro.Schema
 import org.apache.avro.mapreduce.AvroJob
